Reorganize celery tasks
